start of INI revoluation
We have a bajillion runtime options. We should cut them into 3 categories:

* Apache or Nginx setting replacements
* Exact mirrors of php ini settings
* New things that we think should be configurable

This diff tackles the second category. I went through a bunch of the manual pages for ini settings and tried to find the same setting in our hdf. While I was doing that I figured out that exposing some of those to userland would be a security problem, so I introduced the PHP idea of where a setting can be set. The only place I limited it was the `ini_set` call for now, but we can build out more if we support per-dif `php.ini`.

Then while I was doing this, I did lots of stupid things like hooking the wrong variable up to the wrong setter and getter, so I made the compiler make the choice for me by overloading `Bind` a bunch.

So yeah, we got 3 things rolled into this diff (new ini settings, ini setting security, and ini type inference) but I think they jive well together and cutting them apart will be a bit painful.

Reviewed By: @fredemmott

Differential Revision: D1149351

+++ hphp/runtime/base/runtime-option.cpp
@@ -88,8 +88,8 @@ bool RuntimeOption::ThrowInvalidArguments = false;
 bool RuntimeOption::EnableHipHopErrors = true;
 bool RuntimeOption::AssertActive = false;
 bool RuntimeOption::AssertWarning = false;
-int RuntimeOption::NoticeFrequency = 1;
-int RuntimeOption::WarningFrequency = 1;
+int64_t RuntimeOption::NoticeFrequency = 1;
+int64_t RuntimeOption::WarningFrequency = 1;
 int RuntimeOption::RaiseDebuggingFrequency = 1;
 int64_t RuntimeOption::SerializationSizeLimit = StringData::MaxSize;
 int64_t RuntimeOption::StringOffsetLimit = 10 * 1024 * 1024; // 10MB
@@ -307,7 +307,7 @@ int64_t RuntimeOption::MaxRSSPollingCycle = 0;
 int64_t RuntimeOption::DropCacheCycle = 0;
 int64_t RuntimeOption::MaxSQLRowCount = 10000;
 int64_t RuntimeOption::MaxMemcacheKeyCount = 0;
-int RuntimeOption::SocketDefaultTimeout = 5;
+int64_t RuntimeOption::SocketDefaultTimeout = 5;
 bool RuntimeOption::LockCodeMemory = false;
 int RuntimeOption::MaxArrayChain = INT_MAX;
 bool RuntimeOption::WarnOnCollectionToArray = false;
@@ -694,8 +694,8 @@ void RuntimeOption::Load(Hdf &config,
     EnableHipHopErrors = error["EnableHipHopErrors"].getBool(true);
     AssertActive = error["AssertActive"].getBool();
     AssertWarning = error["AssertWarning"].getBool();
-    NoticeFrequency = error["NoticeFrequency"].getInt32(1);
-    WarningFrequency = error["WarningFrequency"].getInt32(1);
+    NoticeFrequency = error["NoticeFrequency"].getInt64(1);
+    WarningFrequency = error["WarningFrequency"].getInt64(1);
   }
   {
     Hdf rlimit = config["ResourceLimit"];
@@ -707,7 +707,7 @@ void RuntimeOption::Load(Hdf &config,
     setResourceLimit(RLIMIT_NOFILE, rlimit, "MaxSocket");
     setResourceLimit(RLIMIT_DATA,   rlimit, "RSS");
     MaxRSS = rlimit["MaxRSS"].getInt64(0);
-    SocketDefaultTimeout = rlimit["SocketDefaultTimeout"].getInt16(5);
+    SocketDefaultTimeout = rlimit["SocketDefaultTimeout"].getInt64(5);
     MaxRSSPollingCycle = rlimit["MaxRSSPollingCycle"].getInt64(0);
     DropCacheCycle = rlimit["DropCacheCycle"].getInt64(0);
     MaxSQLRowCount = rlimit["MaxSQLRowCount"].getInt64(0);
